requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.15
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
gunicorn==21.2.0